Add finished exercise 4.21

+++ controllers/blogs.js
@@ -59,6 +59,24 @@ blogsRouter.post('/', async (request, response, next) => {
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
     const { id } = request.params;
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' });
+    }
+
+    const blog = await Blog.findById(id);
+    if (!blog) {
+      return response.status(404).json({ error: 'blog not found' });
+    }
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+      return response
+        .status(401)
+        .json({ error: 'only the creator can delete this blog' });
+    }
+
     await Blog.findByIdAndDelete(id);
     response.status(204).end();
   } catch (exception) {
